Sorted out a tricky problem with 0-based and 1-based indexing...
The keypoints that we obtain from the OpenCV-based keypoint detectors
are in 0-based coordinate system. Same goes for the Affine dataset's
homography matrices. This caused issues with Matlab implementation
of AlphaGamma descriptor, because there was a 1-pixel offset in the
center coordinates of the patches. Similarly, when plotting the
keypoints in Matlab, the coordinates need to be added 1 pixel, otherwise
they will not be displayed on the correct pixel location...