Returned to the old keypoint identification, via class_id field.
It turned out that it is nigh-impossible to identify keypoints via
their attributes, because ultimately, you can never tell if there
is an angle mismatch or if descriptor extractor chose to overwrite
the angle value.

Therefore, we are back to using the class_id field. We now also
strictly check whether this field is set to -1 before overwriting
it with IDs, to catch keypoint detector/descriptor extractor
implementations that are (ab)using this field.

Consequently, KAZE and AKAZE OpenCV implementations will need to
be modified to not use that field for their purposes, which also
means that for the time being, we are tied to using the modified
OpenCV build...