build_dependencies.sh: disable Lapack support in OpenCV
On Fedora 25 with both OpenBLAS and Atlas installed, the header
from the Atlas gets mixed up for OpenBLAS header, causing build
error. The issue does not appear to have a clear work-around
(aside from giving paths on the command-line, which makes it
non-portable across linux distributions), and as we do not really
need Lapack for our tests, we disable it altogether.