vicos.descriptor.AlphaGamma: fixed a bug in filtering, added weighted distance for float version
The code cleanup introduced regression where the filters were
applied as if they were incremental, when they were not. Now, we
properly apply them to the base image (instead of the previous one)
in the pyramid.

The distance weighting for float version is now implemented by
multiplying corresponding parts of descriptors with weights prior
to computing distance. This seems a bit cleaner than reference
implementation, which applies weights when computing descriptors.
Also, we could compute two sets of distances instead and weight
those, but that would be more demanding in terms of storage
(two potentially large MxN matrices instead of only one)...