Experiment code: reduce memory footprint
Depending on the number of keypoints, the descriptor distance matrix
M can be quite large. And becase we also store it in Mm, which we
later modify to find second-closest match, we consume twice the
amount of memory. As the original M is not really used anymore,
we now modify it directly instead of creating a copy.