Experiments: evaluate_matches: break down descriptor matching into several batches
On some datasets, we get +50k descriptors in both reference and
test image, and the resulting distance matrix consumes a
considerable amount of memory.

In order to keep memory requirements within limits of reasonable,
the processing (descriptor matching) is now split into batches
of K (hardcoded to 1000) test descriptors. If even this proves
problematic, the setting will be exported to the top-level experiment
function for better control...