Experiment: stop using keypoint's class_id
In the very early version of the framework, we used the class_id
field to track which keypoints were discarded by the descriptor
extractor. This, in turn, meant that we needed to ensure that
class_id field was not (ab)used by the keypoint detector, and was
the reason why we were using the modified OpenCV sources.

As we do not make use of the augmented class_id anymore, stop
augmenting it, and stop checking whether the field is unused.
This allows us to use unmodified upstream OpenCV.

NOTE: this change, along with the previous commit, means that
the old cache generated by the experimental framework is potentially
invalid, and should be purged!